-csv2trk broken for very large files
koviz could not convert a 5.4G csv file to trk.

Here is the email:
  Today, we ran into an issue while trying to convert
  some data from csv to trk format.  The example command
  and output follow:

    cmd:  koviz -csv2trk log_mti_a.csv -o log_mti_a.trk
    output: csv2trk [error]: Empty csv file "log_mti_a.csv"
            snap [error]: Aborting csv to trk conversion!

  From our tests, it seems that the size of the CSV file
  is the culprit.  We can convert smaller files fine.
  The file that aborts is 5.4G.  We have to reduce it
  to around 1.1G before it will convert the
  data without the "empty csv file" error.

+++ libsnap/trickmodel.h
@@ -337,7 +337,7 @@ class TrickModelIterator
 
   private:
 
-    int i;
+    quint64 i;
     double _xScaleFactor;
     double _yScaleFactor;
     const TrickModel* _model;
